+++ js/imageeditor/ImageEditor.stories.svelte
@@ -55,7 +55,7 @@
 	}}
 />
 
-<Story
+<!-- <Story
 	name="Image Editor Undo/Redo Interactions"
 	args={{
 		value: {
@@ -108,15 +108,15 @@
 
 		await userEvent.click(canvas.getByLabelText("Redo"));
 	}}
-/>
+/> -->
 
 <Story
 	name="Static Image Display"
 	args={{
 		value: {
 			composite: {
-				path: "",
-				url: "/output-image.png",
+				path: "https://gradio-builds.s3.amazonaws.com/demo-files/ghepardo-primo-piano.jpg",
+				url: "https://gradio-builds.s3.amazonaws.com/demo-files/ghepardo-primo-piano.jpg",
 				size: null,
 				orig_name: null,
 				mime_type: null,