+++ js/imageeditor/shared/tools/index.ts
@@ -1,5 +1,0 @@
-export { default as Tools } from "./Tools.svelte";
-export { default as Brush } from "./Brush.svelte";
-export { default as Crop } from "./Crop.svelte";
-export { default as Sources } from "./Sources.svelte";
-export * from "./types";