+++ js/app/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @self/app
 
+## 1.49.0
+
+### Features
+
+- [#10635](https://github.com/gradio-app/gradio/pull/10635) [`2f68c9d`](https://github.com/gradio-app/gradio/commit/2f68c9d988dcbc53a0b8e53bdb1de49c9c8c65d8) - Refactor and redesign `ImageEditor` component.  Thanks @pngwn!
+
+### Dependency updates
+
+- @gradio/core@0.13.2
+
 ## 1.48.1
 
 ### Dependency updates