+++ js/file/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/file
 
+## 0.12.13
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.7
+- @gradio/atoms@0.15.0
+- @gradio/icons@0.11.0
+- @gradio/upload@0.16.0
+
 ## 0.12.12
 
 ### Dependency updates