+++ js/simpletextbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/simpletextbox
 
+## 0.3.17
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.7
+- @gradio/atoms@0.15.0
+- @gradio/icons@0.11.0
+
 ## 0.3.16
 
 ### Dependency updates