+++ js/textbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/textbox
 
+## 0.10.7
+
+### Dependency updates
+
+- @gradio/statustracker@0.10.7
+- @gradio/atoms@0.15.0
+- @gradio/icons@0.11.0
+
 ## 0.10.6
 
 ### Dependency updates